[IDLE-104] 공고 ContentView들을 BaseFeature모듈로 이동
재사용을 위한 이동입니다.
